[dogstatsd] use a pool of StringSlices instead of re-allocating slices every time.
DRAFT — WIP

The main idea here is to re-use as much pre-allocated memory as possible to
avoid having the GC triggering.

It is mainly removing the slice creation in `parseTags` by reusing slices of String.

Without this code change: 45% of the allocation while receiving count metrics are in `parseTags`.
With this code change: 1% of the allocation (pool instances creation).