Only update telemetry if source was actually deleted
The tag store previously used to always decrease the number of stored
entities even when sourceTags for a source had already been deleted
previously, causing telemetry to slowly decrease to over time to below
zero, which should be impossible. Now, telemetry is only updated if
there actually was a source to be deleted in the first place.