[rtloader] Add ability to set program name to rtloader
Previously, our RTLoader was unable to have a custom program name (via
`Py_SetProgramName`) which made use of libraries like `multiprocessing`
not work since they would try to spawn new processes via the rtloader's
pareent process path.

This change adds an ability to configure this value so that when
a library like `multiprocessing` it can point to the correct interpreter
path and work as intended.

Signed-off-by: Srdjan Grubor <sgnn7@sgnn7.org>