WIP: Preliminary draft RFC for modularizing the `pkg/collector/runner` code
This code should effectively contain almost no functional changes other
than splitting up the `Runner` object into components that can be easier
to understand, expand, and test. The RFC is on the architectural changes
before more time is commited to the change.

**NOTE: This PR is incomplete at this time and tests do not pass since
it's just a RFC on the architectural approach.**